+++ src/seq_jacobi.cpp
@@ -7,16 +7,16 @@
 
 using namespace std;
 
-void seq_jacobi_method(vector<vector<double>> &A, vector<double> &b, double eps, int max_iterations, bool distance_stop, bool debug, vector<double> &res) {
+void seq_jacobi_method(vector<vector<float>> &A, vector<float> &b, float eps, int max_iterations, bool distance_stop, bool debug, vector<float> &res) {
     int k = 0;
 
     int n = A.size();
 
     // initialization variables
-    vector<double> x_k(n, 0);
-    vector<double> x_prev(n, 0);
+    vector<float> x_k(n, 0);
+    vector<float> x_prev(n, 0);
 
-    // getting underling data, return array of double, faster to deal with
+    // getting underling data, return array of float, faster to deal with
     auto A_data = A.data();
     auto b_data = b.data();
     auto x_prev_data = x_prev.data();
@@ -33,7 +33,7 @@ void seq_jacobi_method(vector<vector<double>> &A, vector<double> &b, double eps,
 
         // Jacobi Iteration
         for (int i=0; i<n; i++) {
-            double sigma = 0;
+            float sigma = 0;
             for (int j=0; j<n; j++) {
                 if  ( j!=i )
                     sigma = sigma + A_data[i][j]*x_prev_data[j];
